move sorting icon into different file
