+++ lib/sortable-table.js
@@ -3,6 +3,7 @@ import * as React from 'react';
 import ReactDOM from 'react-dom';
 import move from 'lodash-move';
 import ProgressiveTable from 'react-progressive-table';
+import SortingIcon from './sorting-icon';
 
 /** @memberOf SortableTable */
 type Props = {
@@ -18,25 +19,6 @@ type State = {
     oldIndex: number,
     newIndex: number
   }
-};
-
-const SortingIcon = ({ children, ...rest }: { children?: React.ElementType }) =>
-  React.Children.map(children, (child) => {
-    if (!child || !child.props) return child;
-
-    return React.cloneElement(
-      child,
-      { ...rest, ...child.props },
-      child.props.children
-    );
-  });
-
-SortingIcon.defaultProps = {
-  children: (
-    <span>
-      {'<>'}
-    </span>
-  )
 };
 
 export default class SortableTable extends React.Component<Props, State> {
